+++ src/main/java/com/example/eat_together/domain/cart/controller/SharedCartController.java
@@ -1,9 +1,7 @@
 package com.example.eat_together.domain.cart.controller;
 
 import com.example.eat_together.domain.cart.dto.request.SharedCartItemRequestDto;
-import com.example.eat_together.domain.cart.dto.response.SharedCartItemResponseDto;
 import com.example.eat_together.domain.cart.dto.response.SharedCartResponseDto;
-import com.example.eat_together.domain.cart.entity.SharedCartItem;
 import com.example.eat_together.domain.cart.service.SharedCartService;
 import com.example.eat_together.global.dto.ApiResponse;
 import jakarta.validation.Valid;
@@ -13,8 +11,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -45,7 +41,6 @@ public class SharedCartController {
         return ResponseEntity.ok(ApiResponse.of(response, "공유 장바구니 항목을 조회했습니다."));
     }
 
-
     @PatchMapping("/items/{itemId}")
     public ResponseEntity<ApiResponse<Void>> updateItemQuantity(
             @AuthenticationPrincipal UserDetails userDetails,
@@ -71,14 +66,4 @@ public class SharedCartController {
         return ResponseEntity.ok(ApiResponse.of(null, "공유 장바구니 항목이 삭제되었습니다."));
     }
 
-    @PostMapping("/move-to-cart")
-    public ResponseEntity<ApiResponse<Void>> moveToCart(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable Long roomId,
-            @PathVariable Long storeId
-    ) {
-        Long userId = Long.valueOf(userDetails.getUsername());
-        sharedCartService.moveToCart(userId, roomId);
-        return ResponseEntity.ok(ApiResponse.of(null, "공유 장바구니 항목이 개인 장바구니로 이동되었습니다."));
-    }
 }
